Fix and complete CPP09 ex00/ex01/ex02
- ex00: OCF copy ctor/assign, NaN guard, fix Makefile compile rule and re target
- ex01: simplify to single-pass loop, fix throw type, stack overflow for long long, straggler check
- ex02: fix sortedPairs re-linking bug after recursive sort, bounded binary search, duplicate detection, OCF, Makefile fixes
- Add .gitignore (exclude .claude/, PDF)
- Add ex02/EXPLANATION.md

+++ ex01/RPN.hpp
@@ -7,7 +7,7 @@
 class RPN {
     private:
         std::string _input;
-        std::stack<int> _stack;
+        std::stack<long long> _stack;
     public:
         RPN();
         RPN(const std::string& _inputval);
